esLint and Format

+++ packages/client/src/controllers/AuthController.ts
@@ -1,27 +1,21 @@
-import AuthAPI from '../api/AuthApi/AuthApi';
-import { SignInRequest } from '../store/auth/types';
-
-
+import AuthAPI from '../api/AuthApi/AuthApi'
+import { SignInRequest } from '../store/auth/types'
 
 class AuthController {
-  private readonly api: AuthAPI;
+  private readonly api: AuthAPI
 
   constructor() {
-    this.api = new AuthAPI;
+    this.api = new AuthAPI()
   }
 
   async signin(data: SignInRequest) {
     try {
-      await this.api.signin(data);
-      localStorage.setItem('auth', "userAuthorized")
+      await this.api.signin(data)
+      localStorage.setItem('auth', 'userAuthorized')
     } catch (e: any) {
-      if (
-        e.reason ===
-        'User already in system'
-      ) {
-        localStorage.setItem('auth', "userAuthorized")
-      }
-      else {
+      if (e.reason === 'User already in system') {
+        localStorage.setItem('auth', 'userAuthorized')
+      } else {
         localStorage.removeItem('auth')
         alert('Неверный логин или пароль')
       }
@@ -30,10 +24,11 @@ class AuthController {
 
   async logout() {
     try {
-      await this.api.logout();
-    } catch (e: any) { console.log(e) }
+      await this.api.logout()
+    } catch (e: unknown) {
+      console.log(e)
+    }
   }
 }
 
-export default new AuthController();
-
+export default new AuthController()
